Raster support (#4718)
* adding time series model

* changes

* review changes

* fix lstm

* changes

* changes

* changes

* changes

* changes

* adding support for rasters

* updating models + review changes

* changes

* fix

* changes

* type casting

+++ src/arcgis/learn/models/_layer_learner.py
@@ -559,4 +559,4 @@ class FullyConnectedNetwork(ArcGISModel):
         if self._data._is_classification:
             return (np.array(predictions)==labels).mean()
         else:
-            return float(r2_score(torch.tensor(np.array(predictions)), torch.tensor(np.array(labels))))
+            return float(r2_score(torch.tensor(np.array(predictions, dtype='float64')), torch.tensor(np.array(labels, dtype='float64'))))
